backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
requests
playwright
stripe
schedule
python-dotenv
